requirements.txt
streamlit
pandas
requests
pyarrow
google-cloud-storage
fastparquet